+++ src/utils/helpers/random.helpers.js
@@ -1,0 +1,6 @@
+function randomNumber() {
+    // Returns a random integer from 1 to 100:
+    return Math.floor(Math.random() * 10) + 1;
+}
+
+export{randomNumber};